ts-extras: fix browser entry points pulling in Node.js fs/path modules
readCsvFromTree and readRecordJarFromTree were exported from their
respective *FileHelpers modules, which have unconditional top-level
`import * as fs` / `import * as path`. Bundlers (Vite) statically
trace the require() and emit "Module 'fs' has been externalized"
warnings for every browser consumer.

Fix: move the two FileTree-based helpers into the browser-safe
csvHelpers / recordJarHelpers modules (neither function uses fs or
path), then drop the *FileHelpers imports from the .browser index
files. The Node entry points are unaffected — both functions are
still exported via `export * from './csvHelpers'` /
`export * from './recordJarHelpers'`.
