+++ libraries/ts-extras/src/packlets/csv/csvHelpers.ts
@@ -21,6 +21,7 @@
  */
 
 import { Result, captureResult } from '@fgv/ts-utils';
+import { FileTree } from '@fgv/ts-json-base';
 import { parse } from 'papaparse';
 
 /**
@@ -51,4 +52,23 @@ export function parseCsvString(body: string, options?: CsvOptions): Result<unkno
       ...options
     }).data.slice(1);
   });
+}
+
+/**
+ * Reads a CSV file from a FileTree.
+ * @param fileTree - The FileTree to read from.
+ * @param filePath - Path of the file within the tree.
+ * @param options - optional parameters to control the processing
+ * @returns The parsed CSV data.
+ * @beta
+ */
+export function readCsvFromTree(
+  fileTree: FileTree.FileTree,
+  filePath: string,
+  options?: CsvOptions
+): Result<unknown> {
+  return fileTree
+    .getFile(filePath)
+    .onSuccess((file) => file.getRawContents())
+    .onSuccess((contents) => parseCsvString(contents, options));
 }
