+++ libraries/ts-extras/src/packlets/record-jar/recordJarFileHelpers.ts
@@ -24,7 +24,6 @@ import * as fs from 'fs';
 import * as path from 'path';
 
 import { Result, captureResult } from '@fgv/ts-utils';
-import { FileTree } from '@fgv/ts-json-base';
 import { JarRecord, JarRecordParserOptions, parseRecordJarLines } from './recordJarHelpers';
 
 /**
@@ -47,25 +46,3 @@ export function readRecordJarFileSync(
   });
 }
 
-/**
- * Reads a record-jar file from a FileTree.
- * @param fileTree - The FileTree to read from.
- * @param filePath - Path of the file within the tree.
- * @param options - Optional parser configuration
- * @returns The contents of the file as an array of `Record<string, string>`
- * @see https://datatracker.ietf.org/doc/html/draft-phillips-record-jar-01
- * @public
- */
-export function readRecordJarFromTree(
-  fileTree: FileTree.FileTree,
-  filePath: string,
-  options?: JarRecordParserOptions
-): Result<JarRecord[]> {
-  return fileTree
-    .getFile(filePath)
-    .onSuccess((file) => file.getRawContents())
-    .onSuccess((contents) => {
-      const lines = contents.split(/\r?\n/);
-      return parseRecordJarLines(lines, options);
-    });
-}
