+++ libraries/ts-extras/src/packlets/record-jar/recordJarHelpers.ts
@@ -21,6 +21,7 @@
  */
 
 import { Result, fail, isKeyOf, succeed } from '@fgv/ts-utils';
+import { FileTree } from '@fgv/ts-json-base';
 
 interface IRecordBody {
   body: string;
@@ -251,4 +252,27 @@ class RecordParser {
  */
 export function parseRecordJarLines(lines: string[], options?: JarRecordParserOptions): Result<JarRecord[]> {
   return RecordParser.parse(lines, options);
+}
+
+/**
+ * Reads a record-jar file from a FileTree.
+ * @param fileTree - The FileTree to read from.
+ * @param filePath - Path of the file within the tree.
+ * @param options - Optional parser configuration
+ * @returns The contents of the file as an array of `Record<string, string>`
+ * @see https://datatracker.ietf.org/doc/html/draft-phillips-record-jar-01
+ * @public
+ */
+export function readRecordJarFromTree(
+  fileTree: FileTree.FileTree,
+  filePath: string,
+  options?: JarRecordParserOptions
+): Result<JarRecord[]> {
+  return fileTree
+    .getFile(filePath)
+    .onSuccess((file) => file.getRawContents())
+    .onSuccess((contents) => {
+      const lines = contents.split(/\r?\n/);
+      return parseRecordJarLines(lines, options);
+    });
 }
